Bump version to v1.1, (and change to X.Y scheme)

+++ src/main/java/azurelmao/worldedit/WandPlayerData.java
@@ -3,7 +3,7 @@ package azurelmao.worldedit;
 import java.util.HashMap;
 
 public class WandPlayerData {
-    public static HashMap<String, double[]> primaryPositions = new HashMap<>();
-    public static HashMap<String, double[]> secondaryPositions = new HashMap<>();
+    public static HashMap<String, int[]> primaryPositions = new HashMap<>();
+    public static HashMap<String, int[]> secondaryPositions = new HashMap<>();
     public static HashMap<String, WandClipboard> wandClipboards = new HashMap<>();
 }
